#131 refactor: ProductLike DTO 분리 및 로직 수정
