+++ src/test/java/com/example/testypie/domain/productLike/service/ProductLikeServiceTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.verify;
 import com.example.testypie.domain.product.entity.Product;
 import com.example.testypie.domain.product.repository.ProductRepository;
 import com.example.testypie.domain.product.service.ProductService;
-import com.example.testypie.domain.productLike.dto.ProductLikeResponseDto;
+import com.example.testypie.domain.productLike.dto.response.ProductLikeResponseDto;
 import com.example.testypie.domain.productLike.entity.ProductLike;
 import com.example.testypie.domain.productLike.repository.ProductLikeRepository;
 import com.example.testypie.domain.user.entity.User;
